+++ chat-widget/src/components/webchatcontainerstateful/common/defaultStyles/defaultWebChatStyles.ts
@@ -55,5 +55,5 @@ export const defaultWebChatStyles: StyleOptions = {
         "</3": "💔",
         "<\\3": "💔"
     },
-    uploadMultiple: false
+    sendAttachmentOn: "attach"
 };